requirements.txt
args==0.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clint==0.5.1
colorama==0.4.6
docutils==0.21.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
id==1.5.0
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.21
packaging==24.2
pillow==11.1.0
pim==1.0.1
Pygments==2.19.1
pytz==2025.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==75.8.2
SQLAlchemy==2.0.38
twine==6.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
